refactor(tests): one ToolCall helper, and one code for "not playing"
Twelve fixtures carried a byte-identical private copy of the Call/Ok/Refused/
Code helpers - 240 lines of the same ten. They now pull ToolCall in with
`using static`, so every call site reads exactly as it did and the copies are
gone.

Two contract changes the ponytail pass over the new tests surfaced:

- set_paused and step_frame answered NOT_IN_PLAY_MODE where eight other sites
  answer PLAY_MODE_REQUIRED. Two names for one condition means a client that
  handles one silently fails to recognise the other, so the second name is
  gone rather than kept as an alias. Breaking for anyone branching on it.

- set_tag_and_layer wrapped its tag assignment in catch (UnityException),
  which is how older Unity reports an unknown tag. 6000.3 just takes the
  string, so nothing threw, nothing warned, and the tag did not stick. It
  pre-checks InternalEditorUtility.tags now, the way the layer half of the
  same tool always has.

915 tests, 911 pass, 0 fail, 4 skipped (all pre-existing and documented).
