+++ Tests/Editor/SpriteAndAtlasToolsTests.cs
@@ -1,10 +1,7 @@
 // Copyright (C) KitWright. Licensed under MIT.
 
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using KitWright.Editor.Tools;
-using Newtonsoft.Json.Linq;
 using NUnit.Framework;
 using UnityEditor;
 // The atlas getters this reads (GetPackables, GetPackingSettings, IsIncludeInBuild) are editor-only
@@ -12,6 +9,7 @@ using UnityEditor;
 using UnityEditor.U2D;
 using UnityEngine;
 using UnityEngine.U2D;
+using static KitWright.Editor.Tests.ToolCall;
 
 namespace KitWright.Editor.Tests
 {
@@ -40,25 +38,6 @@ namespace KitWright.Editor.Tests
         {
             AssetDatabase.DeleteAsset(Folder);
         }
-
-        private static JObject Call(string tool, params string[] pairs)
-        {
-            var parameters = new Dictionary<string, string>();
-            for (var i = 0; i + 1 < pairs.Length; i += 2)
-                parameters[pairs[i]] = pairs[i + 1];
-
-            return JObject.Parse(new FunctionInvoker().Invoke(
-                new FunctionCall { FunctionName = tool, Parameters = parameters }));
-        }
-
-        private static JObject Ok(string tool, params string[] pairs)
-        {
-            var answer = Call(tool, pairs);
-            Assert.IsTrue((bool)answer["success"], $"{tool}: {answer}");
-            return answer;
-        }
-
-        private static string Code(string tool, params string[] pairs) => (string)Call(tool, pairs)["code"];
 
         private static TextureImporter Importer(string path) => (TextureImporter)AssetImporter.GetAtPath(path);
