+++ Tests/Editor/ToolCall.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 259e900de1b4488469645616aebbe7eb
